requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
requests==2.32.2
Werkzeug==3.0.6
firebase-admin==6.2.0
gunicorn==23.0.0
PyYAML==6.0.1
pytz==2024.1
openai==1.6.1
pydantic==2.5.3
pydantic-core==2.14.6
httpx>=0.24.1
win10toast==0.9.0; sys_platform == 'win32'
Pillow==10.3.0
SQLAlchemy==2.0.23
PyJWT==2.8.0
Flask-Session==0.5.0
